backend/requirements.txt
flask
flask_sqlalchemy
pymysql
flask-cors
werkzeug
pyjwt
